BASW-147: Implement Membership Status Update
Contribution status change should have caused related memberships to be
updated too. This was not happening because the status was being updated via a
query directly sent to th DB.

Fixed by calling API to update the contribution's status and adding a call to
CRM_Contribute_BAO_Contribution::transitionComponentWithReturnMessage(), which
does the actual work of updating related enities on status change.