DD-64: Implement Mandate Selection Field on Membership Creation
Implemented a buildform hook on the CRM_Financial_Form_Payment form, which is
used by core CiviCRM payment methods to load fields particular to each method,
to load a select combo with existing mandates for the contact. Added option to
create a new mandate from the select combo, which opens a modal form with
mandate creation fields and refreshes mandate select combo on submission, so
the new mandate can be selected.

Refactored some of the original code that handled mandate creation into the
new mandate creation form and out of membership creation post-processing.