ASE-299: Fix Mandate Creation With Different Start Date then Today
When creating a mandate, insertion into the database was done using a manual
INSERT query, and then the civicrm_custom hook was launched so other mandate
fields were generated. This hook requires a third parmeter, which is a list of
the fields and their values in a very specific format, with all the fields'
meta data. This format was not being used when calling the hook.

Fixed by replacing the simple [$field => $value] array with the expected
format.