CGP-1492: Fixing Installation failure on CiviCRM 5.19
When trying to install the extension on CiviCRM 5.19 the following fatal
error appears :

"A fatal error was triggered: One of parameters (value: ) is not of the type Integer"

The error caused by the following line in CiviCMR core :

https://github.com/civicrm/civicrm-core/blob/cbaa9a31f02afda8d2e19610da823fd0c4d28a16/CRM/Utils/Migrate/Import.php#L163

The method in which the line above exist is the one responsible for
processing any option value defined in the extension XML files, the
method first check if the option group for which the option value belong
is defined in the XML file itself and if not it try to look it up in
already defined CiviCRM option groups, this line is responsible for the
latter and it passes "$optionValueXML->option_group_name" to
getFieldValue() method but since $optionValueXML is a SimpleXML object
and accessing any SimpleXML object element will also return another
SimpleXML object, the getFieldValue() method will fail.

The ideal solution is to update CiviCRM core and cast the result of
"$optionValueXML->option_group_name" to string which is what I am
planning to do, but for the time being I just fixed this issue by
defining any option group that belong to any option value in the XML
file itself so the core code above will not have to reach to the line
that cause the issue and since CiviCRM core already prevent creating any
option group that is already defined then adding these option groups to
the XML file won't have any side effects.