CGP-1472: Implement Validation of Mandate Field on Membership Creation
Membership creation form was allowing mandate field to be empty when Direct
Debit was selected as payment method. This caused a fatal error to be thrown
when saving the membership.

Fixed by validating mandate is not empty if direct debit payment method has
been chosen. Also added some js logic to leave the membersip creation form in
a consistent state after a failed validation, as contact_id and the mandate
field were being lost.