SPAB-976: Check if Membership Exists When Collecting Data
When collecting payment plan data for notifications, if a membership was
deleted, but the contribution remained, an exception was thrown when trying to
obtain the information for the next payment, as the query expected an integer,
but a null value was being given (corresponding to the missing membership).
The message for the exception read "One of parameters  (value: ) is not of the
type Integer".

Fixed by checking if the membership exists when pulling its info with the API
call, and throwing an exception with a less cryptic message. Also added
exception handling to other classes that used the same class to collect and
build the data for messages.