IO-111: Fix Search Form "To" Field
When choosing a date for the "To" field, any creation dates on that same date
would not be included in the search results. This happened because the search
was not taking into account created date is a datetime field, thus any date
with a time over 00:00:00 would be out of the range.

Fixed by forcing "To" date to include all dates before '23:59:59'.