IO-134: Fix Scenario Where First Instalment is After Second
Under some configurations, if setting to force second instalment on second
month is enabled, it can happen that the receive date for the second
instalment is calculated to be before the first intalment. We need to fix this
so if this happens, the receive date of the second instalment is set equal to
the first.