MAE-494: Fix payment batch submission after CiviCRM 5.35 upgrade
Manual direct debit is using the core method "transitionComponentWithReturnMessage" that is in the Contribution BAO class, though this method is now removed in core (see https://github.com/civicrm/civicrm-core/pull/18961 for more info), and thus it result in failure when submitting the payment batch.

Here I replaced the calls for this method with "transitionComponents",
given that the former is basicly calling it but just return some extra
feedback messages, and that these returned feedback messages are not
even used or showed back to user in this extension, so it is safe to
just call "transitionComponents" instead.