MAE-494: Fix PHP warnings
When using the API v3 `getvalue` action on any entity, it will return
the return value directly, but some places in direct debit importer
assumed that it would return the value indexed by the `result` key.

Here I am changing that so results from this API action are used directly.