MAE-579: Remove transitionComponents() method call
When users submit Direct debit payment batch, we record a payment against every contribution
in the submitted batch, this was achived by doing two things:

A- Calling CiviCRM Contribution.Create API to update the contribution with "Completed" status.

B- Calling "CRM_Contribute_BAO_Contribution::transitionComponents()" method that handles updating the entities that are related to the contribution, which is in Direct Debit case is always "Memberships".

This behaviour was implemented long time ago where many things were not clear back then, but I think the idea was to trying to mimic the way CiviCRM core batch submission work:

https://github.com/civicrm/civicrm-core/blob/386ef1451b34e5d324b46688322e38c231e255b6/CRM/Contribute/Form/Task/Batch.php#L208-L214

But it seem now that CiviCRM deprecated CRM_Contribute_BAO_Contribution::transitionComponents(), and their recommendation is to use Payment.Create API to complete payments and update the related
entities instead of calling transitionComponents().

So I initialy tried to replace transitionComponents() with Payment.Create API, but it seems that Payment.Create API does not pass hooks which is the case with transitionComponents() which results
in calling many hooks many times are resulted in creating many unnecessary records and trying to update many entites more than one time which is not good performance-wise, here is an example:

A- transitionComponents() queries (hence that I only put the query type and table name because I am just trying to show the difference in quantity of performed queries):

- UPDATE  `civicrm_contribution`
- INSERT INTO `civicrm_financial_trxn`
- INSERT INTO `civicrm_entity_financial_trxn`
- UPDATE civicrm_financial_item
- INSERT INTO `civicrm_entity_financial_trxn`
- UPDATE  `civicrm_line_item`
- UPDATE  `civicrm_contribution_recur`
- INSERT INTO `civicrm_activity`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`
- UPDATE  `civicrm_membership`
- UPDATE  `civicrm_membership`
- UPDATE  `civicrm_membership`
- INSERT INTO `civicrm_membership_log`

 B- Payment.Create API queries:

- INSERT INTO `civicrm_financial_trxn`
- INSERT INTO `civicrm_entity_financial_trxn`
- INSERT INTO `civicrm_entity_financial_trxn`
- UPDATE  `civicrm_contribution_recur`
- UPDATE  `civicrm_membership`
- INSERT INTO `civicrm_activity`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`
- INSERT INTO `civicrm_membership_log`
- UPDATE  `civicrm_membership`
- INSERT INTO `civicrm_membership_log`
- INSERT INTO `civicrm_activity` (
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`
- INSERT INTO `civicrm_activity`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`
- UPDATE  `civicrm_contribution`
- UPDATE  `civicrm_contribution_recur`
- INSERT INTO `civicrm_activity`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`
- UPDATE  `civicrm_contribution`
- UPDATE  `civicrm_line_item`
- UPDATE  `civicrm_activity`
- INSERT INTO `civicrm_log`
- UPDATE  `civicrm_contribution`
- INSERT INTO `civicrm_activity`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_activity_contact`
- INSERT INTO `civicrm_log`

So as you can see, there are more many more queires when call Payment.Create API and all of these extra queires are not really needed, they are just the effect of not passing hooks which result
in calling certain hooks when there is no need to.

Then I realized that we probably do not need to call transitionComponents() to begin with and all we need to do is only updating the contribution status using Contribution.Create API, the reason for
that is we really don't want to alter related entites like memberships when we complete direct debit payments because Membershipextras and Manual direct debit extensions are differenet from CiviCRM core
when it comes to the relation between the memberships and payments, we kinda seperated the two concepts and thus when you create a Membership with Payment plan all the contributions will be pending
but the membership will be "Active", where in CiviCRM core if you create a membership with pending contribution then the membership will also be pending and it will only become active if you complete
the pending contribution, so the idea of calling transitionComponents() for each recorded contribution after submitting a direct debit batch is not really needed because we don't really need to update
any related entity (which is as we said the memberships in this case).

So what I did here is just removing the call to transitionComponents() and just keeping the API call that update the contribution status to Completed, which should be enough in our case. And just for
the record, ehre are the queires that are called when updating the contribution status with the Contribution.Create API:

- UPDATE civicrm_contribution
- INSERT civicrm_financial_trxn
- INSERT civicrm_entity_financial_trxn
- UPDATE civicrm_financial_item
- INSERT civicrm_entity_financial_trxn
- UPDATE civicrm_line_item
- UPDATE civicrm_contribution_recur
- INSERT civicrm_activity
- INSERT civicrm_activity_contact
- INSERT civicrm_activity_contact
- INSERT civicrm_log

which is good enough since it only do minimal amount of needed queries while ensuring that accounting records are created correctly.